requirement.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
gTTS==2.5.4
h11==0.16.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
outcome==1.3.0.post0
pyfiglet==1.0.4
pygame==2.6.1
Pygments==2.19.2
PySocks==1.7.1
python-nmap==0.7.1
pyttsx3==2.99
requests==2.32.5
rich==14.1.0
scapy==2.6.1
selenium==4.35.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
termcolor==3.1.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
